[FlexNG] Fragmentation for stretched items
As suggested in CL:3256656, refactor ApplyStretchAlignmentToChild()
to use BuildSpaceForLayout. To accomplish this, pass the final line
cross size into BuildSpaceForLayout during the final layout pass for
stretched items.

By doing this we can remove ApplyStretchAlignmentToChild(), and call
into LayoutWithBlockFragmentation() for stretched items if applicable
(since that will also call into BuildSpaceForLayout).

A new wpt test was added to test basic fragmentation for stretched flex
items.

virtual/layout_ng_flex_frag/fast/multicol/flexbox/doubly-nested-with-zero-width-flexbox-and-forced-break-crash.html
is now being skipped due to a timeout. Seems to be caused by the
forced break. This test was not timing out before since the
test was previously falling back to legacy. This will be
investigated in a follow-up.

Bug: 660611
Change-Id: Ifefbe1480c7134c5fbf215cd1e03a8e79e45f1ff